packages renamed to lower case
