+++ app/src/main/java/com/ecommerce/molbhaav/activities/HomePage.java
@@ -1,26 +1,16 @@
-package com.ecommerce.molbhaav.Activities;
+package com.ecommerce.molbhaav.activities;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import android.app.Activity;
-import android.os.Bundle;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.widget.LinearLayout;
 
 
-import com.ecommerce.molbhaav.Adapter.MyAdapter;
-import com.ecommerce.molbhaav.Controller.ItemData;
+import com.ecommerce.molbhaav.adapter.MyAdapter;
+import com.ecommerce.molbhaav.controller.ItemData;
 import com.ecommerce.molbhaav.R;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import okhttp3.OkHttpClient;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
 
 public class HomePage extends Activity {
     @Override
