+++ app/src/main/java/com/ecommerce/molbhaav/activities/MainActivity.java
@@ -1,11 +1,10 @@
-package com.ecommerce.molbhaav.Activities;
+package com.ecommerce.molbhaav.activities;
 
 import android.content.Intent;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-import com.ecommerce.molbhaav.Activities.LoginActivity;
 import com.ecommerce.molbhaav.R;
 
 public class MainActivity extends AppCompatActivity {
