+++ app/src/main/java/com/ecommerce/molbhaav/activities/Registration.java
@@ -1,11 +1,9 @@
-package com.ecommerce.molbhaav.Activities;
+package com.ecommerce.molbhaav.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.text.method.LinkMovementMethod;
 import android.view.View;
 import android.widget.Button;
@@ -14,10 +12,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 
-import com.ecommerce.molbhaav.Controller.IApi;
+import com.ecommerce.molbhaav.controller.IApi;
 import com.ecommerce.molbhaav.R;
-import com.ecommerce.molbhaav.Request.SignInRequest;
-import com.ecommerce.molbhaav.Response.SignInResponse;
+import com.ecommerce.molbhaav.request.SignInRequest;
+import com.ecommerce.molbhaav.response.SignInResponse;
 
 import okhttp3.OkHttpClient;
 import retrofit2.Call;
