+++ app/src/main/java/com/ecommerce/molbhaav/request/SignInRequest.java
@@ -1,11 +1,12 @@
-package com.ecommerce.molbhaav.Request;
+package com.ecommerce.molbhaav.request;
+
 
 
 
 import com.google.gson.annotations.SerializedName;
 
 
-public class LoginRequest{
+public class SignInRequest{
 
     @SerializedName("password")
     private String password;
@@ -18,9 +19,6 @@ public class LoginRequest{
 
     @SerializedName("emailId")
     private String emailId;
-
-    @SerializedName("userId")
-    private String userId;
 
     public void setPassword(String password){
         this.password = password;
@@ -54,23 +52,14 @@ public class LoginRequest{
         return emailId;
     }
 
-    public void setUserId(String userId){
-        this.userId = userId;
-    }
-
-    public String getUserId(){
-        return userId;
-    }
-
     @Override
     public String toString(){
         return
-                "LoginRequest{" +
+                "Response{" +
                         "password = '" + password + '\'' +
                         ",address = '" + address + '\'' +
                         ",name = '" + name + '\'' +
                         ",emailId = '" + emailId + '\'' +
-                        ",userId = '" + userId + '\'' +
                         "}";
     }
 }
