+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activities.LoginActivity">
+    tools:context=".activities.LoginActivity">
     <TextView
         android:id="@+id/loginscrn"
         android:layout_width="wrap_content"
